backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==5.0.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
limits==5.6.0
motor==3.3.2
packaging==25.0
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.6.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.1
